% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_proj}
\alias{create_proj}
\title{Create Project}
\usage{
create_proj(path, dir)
}
\arguments{
\item{path}{Path to copy to.}

\item{dir}{Internal directory containing project.}
}
\description{
Create a project, technically just copies files.
}
\keyword{internal}
