% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{edit_package}
\alias{edit_package}
\title{Edit Package}
\usage{
edit_package(path)
}
\arguments{
\item{path}{Path to the package}
}
\description{
Edits the package name after files have been created.
}
\keyword{internal}
