% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{generator}
\alias{generator}
\alias{create_basic}
\alias{create_bootstrap}
\alias{create_vue}
\title{Create Ambiorix Projects}
\usage{
create_basic(path)

create_bootstrap(path)

create_vue(path)
}
\arguments{
\item{path}{Path where to create ambiorix application.}
}
\description{
Create the basic structure for Ambiorix projects.
}
\section{Projects}{

\itemize{
\item \code{create_basic}: A basic template.
\item \code{create_bootstrap}: A bootstrap 4 template.
\item \code{create_vue}: A template using vue.
}
}

