% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{templates}
\alias{templates}
\alias{add_template_basic}
\alias{add_template_bootstrap}
\title{Add a template file}
\usage{
add_template_basic(name)

add_template_bootstrap(name)
}
\arguments{
\item{name}{Name of the file, without extension.}
}
\description{
Convenience function to create new templates in the \code{templates} directory.
}
\section{Functions}{

\itemize{
\item \code{add_template_basic}: Add a basic template.
\item \code{add_template_bootstrap}: Add a basic template for bootstrap 4.
}
}

