% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{generator}
\alias{generator}
\alias{create_basic}
\alias{create_bootstrap}
\alias{create_vue}
\alias{create_package}
\alias{create_backend}
\title{Create Ambiorix Projects}
\usage{
create_basic(path)

create_bootstrap(path)

create_vue(path)

create_package(path)

create_backend(path, use_box = TRUE)
}
\arguments{
\item{path}{Path where to create ambiorix application.}

\item{use_box}{Logical. Whether to use \href{https://klmr.me/box/}{\code{{box}}} for a
modular app structure.
Default is \code{TRUE}.
If \code{FALSE}, the template will be same as if you used \code{\link[=create_package]{create_package()}}.}
}
\description{
Create the basic structure for Ambiorix projects.
}
\section{Projects}{

\itemize{
\item \code{create_package}: A basic app in R package form (recommended).
\item \code{create_basic}: A basic template.
\item \code{create_bootstrap}: A bootstrap 5 template.
\item \code{create_vue}: A template using vue.
}
}

