% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{generator}
\alias{generator}
\alias{create_basic}
\alias{create_bootstrap}
\alias{create_vue}
\alias{create_package}
\alias{create_modular}
\title{Create Ambiorix Projects}
\usage{
create_basic(path)

create_bootstrap(path)

create_vue(path)

create_package(path)

create_modular(path, project_type = "backend")
}
\arguments{
\item{path}{String. Path to create template in.
If \code{FALSE}, the template will be same as if you used \code{\link[=create_package]{create_package()}}.}

\item{project_type}{String. The project type. Either "backend" (default) or
"frontend".}
}
\description{
Create the basic structure for Ambiorix projects.

This approach creates a modular & nested app template which uses
\href{https://klmr.me/box/}{{box}} to load & unload modules.
}
\section{Projects}{

\itemize{
\item \code{create_package}: A basic app in R package form (recommended).
\item \code{create_basic}: A basic template.
\item \code{create_bootstrap}: A bootstrap 5 template.
\item \code{create_vue}: A template using vue.
}
}

\examples{
\dontrun{
create_modular(path = ".", project_type = "backend")
create_modular(path = ".", project_type = "frontend")
}
}
